updating assignment 4
